requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
requests>=2.31.0
beautifulsoup4>=4.12.0
fastapi>=0.115.0
jinja2>=3.1.0
python-multipart>=0.0.9
uvicorn>=0.30.0
openpyxl>=3.1.0
